requirements.txt
torch
numpy
gym
tensorboard
tensorboardX
xando
pytest
scipy
rlprop